Properly purge unmanaged files in the objects/* subdirectories
The purge_unmanaged_object_files parameter is applied to the objects
directory with recursion and force flags enabled, however this behaviour
seems to be turned back off by the file resources for each subdirectory
and as such unmanaged files are not really purged.

This commit copies the purge parameters from the parent objects directory
onto each child directory to enable the intended behaviour.