Make service name and packages order agnostic
Puppet versions < 4.0 do not resolve correctly scoped variables in some
cases. It is a lot more order dependent. To avoid that, this commit uses
a service with a static title and a dynamic name => attribute.

For the packages, as the title is an array, it is not possible to use a
name => attribute. To have a static resource I have then chosen to use
an anchor, which is dependent on the packages.