apply_scheduleddowntimes folder not necessary
The `target_dir` was set with a wrong folder name. It was not matching the folder in `config.pp`. Here a 's' was missing `apply_scheduleddowntimes` instead of `applys_scheduleddowntimes`!

Adding the object `apply_scheduleddowntimes` the contributer added a new folder to store them.
Understanding that this is an interesting idea, all other apply objects use a general `applys` folder.

Changing this behavior should be generalized for all or none.