Clean up database documentation and add tests
Without this work, the documentation around setting up an initial server
is unclear.  This patch addresses this by fleshing out the documentation
after reviewing the code carefully to discover what the actual behavior
is.  Tests are also added here to ensure that what the documentation
says is true, is actually true.